Agrega verificaciones
Se agregaron verificaciones para las opciones Agregar cuenta y Eliminar cuenta.

+++ main.py
@@ -1,13 +1,23 @@
 #!/bin/python3
 
+"""
+Script para almacenar cuentas.
+Opcion de copiar nombre de cuenta y clave al clipboard.
+
+TODO: crear una GUI.
+
+"""
+
 import os
 import sys
 import pyperclip as pc
 from cryptography.fernet import Fernet
+from getpass import getpass
 
 def opcion1(key):
     os.system("clear")
     flag = False
+    psFlag = False
     account = input("Ingrese nombre de cuenta: ")
     PSSWD = open('.passwd.txt')
     for line in PSSWD:
@@ -20,14 +30,22 @@ def opcion1(key):
     if flag != True:
         user = input("Ingrese nombre de usuario de la cuenta: ")
         PSSWD = open('.passwd.txt', 'a')
-        passwd = input("Ingrese password: ")
-        passwd = passwd.encode()
-        f = Fernet(key)
-        passwd = f.encrypt(passwd)
-        passwd = passwd.decode()
-        newLine = account + ' ' + user + ' ' + passwd + '\n'
-        PSSWD.write(newLine)
+        passwd = getpass("Ingrese password:")
+        passwd2 = getpass("Ingrese de nuevo su password:")
+        if passwd == passwd2:
+            passwd = passwd.encode()
+            f = Fernet(key)
+            passwd = f.encrypt(passwd)
+            passwd = passwd.decode()
+            psFlag = True
+        else:
+            input("Error: passwords no coinciden. Intente de nuevo...")
+            opcion1(key)
+        if psFlag:
+            newLine = account + ' ' + user + ' ' + passwd + '\n'
+            PSSWD.write(newLine)
         PSSWD.close()
+        print(f'Cuenta {account} creada con exito.')
     input("Pulsa enter para continuar.")
     inicio(key)
 
@@ -55,7 +73,7 @@ def opcion3(key):
             opcion = int(input("Copiar usuario (1) o password (2)? "))
             if opcion == 1:
                 pc.copy(line[1])
-                print("Usuario para cuenta " + account + " copiada al clipboard.")
+                print(f'Usuario para cuenta {account} copiada al clipboard.')
                 flag = True
             elif opcion == 2:
                 passwd = line[2]
@@ -64,31 +82,55 @@ def opcion3(key):
                 passwd = f.decrypt(passwd)
                 passwd = passwd.decode()
                 pc.copy(passwd)
-                print("Password para cuenta " + account + " copiada al clipboard.")
+                print(f'Password para cuenta {account} copiada al clipboard.')
                 flag = True
             else:
                 input("Error: opcion incorrecta. Intente de nuevo...")
                 opcion3(key)
     if flag != True:
-        print("No existe la cuenta " + account)
+        print(f'Error: no existe la cuenta {account}')
     input("Pulsa enter para continuar.")
     inicio(key)
 
 def opcion4(key):
     os.system("clear")
+    flag = False
     account = input("Cuenta a eliminar: ")
+    verification = input(f'Seguro que quiere eliminar la cuenta {account}? [y/n]: ')
+    if verification == 'n':
+        input("Pulse enter para continuar.")
+        inicio(key)
     #TODO: verificar password de cuenta antes de eliminarla
-    #passwd = input("Ingrese la password de esa cuenta: ")
+    passwdInput = getpass("Ingrese la password de esa cuenta:")
     with open('.passwd.txt', 'r+') as PSSWD:
         new_PSSWD = PSSWD.readlines()
         PSSWD.seek(0)
         for line in new_PSSWD:
-            line2 = line
             line2 = line.strip().split()
-            if account != line2[0]:
-                PSSWD.write(line)
-        PSSWD.truncate()
-        print("Cuenta " + account + " eliminada con exito.")
+            if line2[0] == account:
+                passwdAc = line2[2]
+                passwdAc = passwdAc.encode()
+                f = Fernet(key)
+                passwdAc = f.decrypt(passwdAc)
+                passwdAc = passwdAc.decode()
+                if (passwdInput == passwdAc):
+                    flag = True
+                    break
+                else:
+                    break
+        PSSWD.close()
+        if flag and verification == 'y':
+            with open('.passwd.txt', 'r+') as PSSWD:
+                new_PSSWD2 = PSSWD.readlines()
+                PSSWD.seek(0)
+                for line in new_PSSWD2:
+                    line2 = line.strip().split()
+                    if account != line2[0]:
+                        PSSWD.write(line)
+                PSSWD.truncate()
+                print(f'Cuenta {account} eliminada con exito.')
+        else:
+            print("Error: password incorrecta. Intente de nuevo...")
     input("Pulsa enter para continuar.")
     inicio(key)
 
@@ -99,7 +141,7 @@ def opcion5():
 def inicio(key):
     os.system("clear")
     print("Inicio.")
-    print("Selecciona una opcion:")
+    print("Seleccione una opcion:")
     print("1: agregar cuenta.")
     print("2: ver cuentas.")
     print("3: copiar password.")
